+++ index.js
@@ -6,3 +6,4 @@ export {uuid};
 export {changeFavicon} from './src/changeFavicon';
 export {getBestLocale, getBrowserLocales} from './src/locales';
 export {URI};
+export {sanitizeFileName} from './src/sanitizeFileName';